Refactor(coverage): Consolidate unit tests and fix urllib import
Consolidated coverage_android_test.py and component_mappings_test.py into a single coverage_unit_test.py file for better organization and maintainability.

Also, fixed a NameError in _GetComponentMappings by explicitly importing the urllib module, which was previously missing when urllib.request.urlopen was called in the fallback logic.
